style: 주석 수정
